Added StreamSound support
