+++ addons/source-python/packages/source-python/engines/sound/base.py
@@ -1,19 +1,21 @@
-# ../engines/sound.py
+# ../engines/sound/base.py
 
-"""Provides access to the Sound interface."""
+"""Provides access to the _BaseSound interface."""
 
 # =============================================================================
 # >> IMPORTS
 # =============================================================================
-# Python Imports
-#   Enum
-from enum import Enum
-
 # Source.Python Imports
 #   Core
 from core import AutoUnload
 #   Engines
-from engines import engines_logger
+from engines.sound import Attenuation
+from engines.sound import Channel
+from engines.sound import engine_sound
+from engines.sound import Pitch
+from engines.sound import SOUND_FROM_WORLD
+from engines.sound import SoundFlags
+from engines.sound import VOL_NORM
 #   Entities
 from entities.constants import INVALID_ENTITY_INDEX
 #   Filters
@@ -21,72 +23,13 @@ from filters.recipients import RecipientFilter
 #   Mathlib
 from mathlib import NULL_VECTOR
 #   Stringtables
-from stringtables import string_tables
 from stringtables.downloads import Downloadables
-
-
-# =============================================================================
-# >> FORWARD IMPORTS
-# =============================================================================
-# Source.Python Imports
-#   Engines
-from _engines._sound import Channel
-from _engines._sound import VOL_NORM
-from _engines._sound import ATTN_NONE
-from _engines._sound import ATTN_NORM
-from _engines._sound import ATTN_IDLE
-from _engines._sound import ATTN_STATIC
-from _engines._sound import ATTN_RICOCHET
-from _engines._sound import ATTN_GUNFIRE
-from _engines._sound import MAX_ATTENUATION
-from _engines._sound import SoundFlags
-from _engines._sound import Pitch
-from _engines._sound import SOUND_FROM_LOCAL_PLAYER
-from _engines._sound import SOUND_FROM_WORLD
-from _engines._sound import engine_sound
-
-
-# =============================================================================
-# >> ALL DECLARATION
-# =============================================================================
-__all__ = ('Attenuation',
-           'Channel',
-           'Pitch',
-           'SOUND_FROM_LOCAL_PLAYER',
-           'SOUND_FROM_WORLD',
-           'Sound',
-           'SoundFlags',
-           'VOL_NORM',
-           'engine_sound',
-           )
-
-
-# =============================================================================
-# >> GLOBAL VARIABLES
-# =============================================================================
-# Get the sp.engines.sound logger
-engines_sound_logger = engines_logger.sound
-
-
-# =============================================================================
-# >> ENUMERATORS
-# =============================================================================
-class Attenuation(float, Enum):
-    """Attenuation values wrapper enumerator."""
-
-    NONE = ATTN_NONE
-    NORMAL = ATTN_NORM
-    IDLE = ATTN_IDLE
-    STATIC = ATTN_STATIC
-    RICOCHET = ATTN_RICOCHET
-    GUNFIRE = ATTN_GUNFIRE
-    MAXIMUM = MAX_ATTENUATION
 
 
 # =============================================================================
 # >> CLASSES
 # =============================================================================
-class Sound(AutoUnload):
+class _BaseSound(AutoUnload):
     """Class used to interact with a specific sound file."""
 
     # Set the base _downloads attribute to know whether
@@ -138,11 +81,7 @@ class Sound(AutoUnload):
             self.precache()
 
         # Play the sound
-        engine_sound.emit_sound(
-            recipients, self.index, self.channel, self.sample,
-            self.volume, self.attenuation, self.flags, self.pitch,
-            self.origin, self.direction, self.origins,
-            self.update_positions, self.sound_time, self.speaker_entity)
+        self._play(recipients)
 
     def stop(self, index=None, channel=None):
         """Stop a sound from being played."""
@@ -159,16 +98,24 @@ class Sound(AutoUnload):
             channel = self.channel
 
         # Stop the sound
-        engine_sound.stop_sound(index, channel, self.sample)
+        self._stop(index, channel)
+
+    def _play(self, recipients):
+        """Play the sound (internal)."""
+        raise NotImplementedError
+
+    def _stop(self, index, channel):
+        """Stop a sound from being played (internal)."""
+        raise NotImplementedError
 
     def precache(self):
         """Precache the sample."""
-        engine_sound.precache_sound(self.sample)
+        raise NotImplementedError
 
     @property
     def is_precached(self):
         """Return whether or not the sample is precached."""
-        return self.sample in string_tables.soundprecache
+        raise NotImplementedError
 
     @property
     def sample(self):
